requirements.txt
streamlit==1.50.0
python-dotenv==1.2.1
PyPDF2==3.0.1

langchain==0.3.28
langchain-core==0.3.83
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-text-splitters==0.3.11

openai==2.26.0
tiktoken==0.12.0
faiss-cpu==1.13.0
fastapi==0.118.2
uvicorn==0.37.0
